[refactor] CalculatorTest와 ValidatorTest 분리
Calculator에서 Validator를 분리해내면서 Test도 분리한다.
